change count from int to NSUInteger

+++ QiniuSDK/Http/QNStats.h
@@ -44,7 +44,7 @@
 - (NSString *) getSIP;
 - (NSString *) getNetType;
 
-@property int count;
+@property NSUInteger count;
 
 @end
